Update projectscript.R
Created dataframe of the RPKM gene expression values of the 3 proteins in anaerbiosis conditions from phytozome (GSE42035). Tried to make PCA to investigate variation between gene expression in light oxic and dark anoxic 6 hours - but couldn't do it. The PC1 and PC2 value of a gene across the different conditions is the same and I can't work out why yet. Next step is to do DeSeq2.