Filtered for top 1000 variance genes
Filtered for the top 1000 variance genes + added back the genes of interest as they were not within the top 1000 variance genes. I'm wondering if I was too strict with my cutoffs, as the variance values of the top 50 coexpressed genes in phytozome are well below the variance of the top 1000 variance genes. Could include maybe the top 5000 variance genes but this would be too computationally intensive for me.